Added readme to light dark demo.
