+++ samples/dashboard/demos/dashboard-light-dark-theme/demo.css
@@ -3,12 +3,6 @@
 @import "https://code.highcharts.com/css/dashboard/gui.css";
 @import "https://code.highcharts.com/css/dashboard/lightDarkTheme.css";
 
-#datagrid {
-    height: 180px;
-    width: 100%;
-    margin: 0 auto;
-}
-
 #csv {
     display: none;
 }
