+++ samples/dashboard/demos/dashboard-light-dark-theme/demo.html
@@ -1,7 +1,9 @@
 <script src="https://code.highcharts.com/dashboard.js"></script>
-<script src="https://code.highcharts.com/highcharts.js"></script>
 <script src="https://code.highcharts.com/datagrid.js"></script>
-<div id="container" class="highcharts-dashboard"></div>
+<script src="https://code.highcharts.com/highcharts.js"></script>
+<script src="https://code.highcharts.com/modules/dashboard-plugin.js"></script>
+
+<div id="container" class="highcharts-dashboard">
 
 <pre id="csv">Food,Vitamin A
 Beef Liver,6421
