+++ samples/dashboard/demos/dashboard-light-dark-theme/readme.md
@@ -1,0 +1,5 @@
+# Light dark theme detection
+This demo contains a CSS module called `lightDarkTheme`. This allows to style off the dashboard based on the theme selected on the user's machine.<br>
+If your dashboard contains Highcharts charts, you have to enable the styled mode.
+
+To test it, please change the default theme in your computer options.
